+++ src/convertFormats.hpp
@@ -16,12 +16,12 @@ CSC<ValueType> COOToCSC(const COO<ValueType> &);
 template <typename ValueType>
 size_t findEllCols(const COO<ValueType> &, const size_t);
 template <typename ValueType>
-std::unique_ptr<size_t[]> findColBlockIdx(const COO<ValueType> &, const size_t,
-                                          const size_t, const size_t);
+std::unique_ptr<int[]> findColBlockIdx(const COO<ValueType> &, const size_t,
+                                       const size_t, const size_t);
 template <typename ValueType>
 std::unique_ptr<ValueType[]> findVals(const COO<ValueType> &, const size_t,
                                       const size_t, const size_t,
-                                      const std::unique_ptr<size_t[]> &);
+                                      const std::unique_ptr<int[]> &);
 template <typename ValueType>
 BELL<ValueType> COOToBELL(const COO<ValueType> &, size_t);
